update checks & output for usernames
